Renamed IWebSocket.h to IWebSocketResource.h.
