+++ vnext/Desktop.IntegrationTests/WebSocketIntegrationTest.cpp
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 #include <CppUnitTest.h>
-#include <IWebSocket.h>
+#include <IWebSocketResource.h>
 #include <Test/WebSocketServer.h>
 
 #include <condition_variable>
