+++ vnext/ReactWindowsCore/IWebSocketResource.h
